requirements/base.txt
# Django
# --------------------
Django==5.0.1
psycopg2==2.9.9
pillow==10.2.0
numpy==1.26.3
djangorestframework==3.14.0
drf-spectacular==0.27.1
django-filter==23.5
pre-commit==3.6.0
matplotlib==3.8.2


